+++ web/src/Layout.tsx
@@ -1,7 +1,6 @@
 import type { PropsWithChildren } from "react";
 
-import Footer from "./components/Footer";
-import Header from "./components/Header";
+import { Footer, Header } from "./components";
 
 function Layout({ children }: PropsWithChildren) {
   return (
