+++ web/src/components/index.ts
@@ -1,0 +1,3 @@
+export { Footer } from "./Footer/Footer";
+export { Header } from "./Header/Header";
+export { SummaryCard } from "./SummaryCard/SummaryCard";
