--- web/src/pages/CreatorPage.tsx
@@ -1,7 +1,0 @@
-import { wrapperStyle } from "./HomePage.css";
-
-function CreatorPage() {
-  return <div className={wrapperStyle}>크리에이터 페이지</div>;
-}
-
-export default CreatorPage;
