+++ web/src/pages/CreatorPage/CreatorPage.css.ts
@@ -1,6 +1,6 @@
 import { style } from "@vanilla-extract/css";
 
-export const wrapperStyle = style({
+export const container = style({
   display: "flex",
   flexDirection: "column",
   alignItems: "center",
@@ -10,7 +10,4 @@ export const wrapperStyle = style({
   height: "calc(100vh - 80px - 100px)",
 
   backgroundColor: "#F1F1F1",
-
-  fontSize: "20px",
-  fontWeight: "bold",
 });
