+++ web/src/pages/CreatorPage/CreatorPage.tsx
@@ -1,0 +1,5 @@
+import * as styles from "./CreatorPage.css";
+
+export function CreatorPage() {
+  return <div className={styles.container}>크리에이터 페이지</div>;
+}
