--- web/src/pages/KingmojangPage.tsx
@@ -1,7 +1,0 @@
-import { wrapperStyle } from "./HomePage.css";
-
-function KingmojangPage() {
-  return <div className={wrapperStyle}>킹모장 페이지</div>;
-}
-
-export default KingmojangPage;
