+++ web/src/pages/KingmojangPage/KingmojangPage.tsx
@@ -1,0 +1,5 @@
+import * as styles from "./KingmojangPage.css";
+
+export function KingmojangPage() {
+  return <div className={styles.container}>킹모장 페이지</div>;
+}
