+++ web/src/pages/index.ts
@@ -1,0 +1,3 @@
+export { CreatorPage } from "./CreatorPage/CreatorPage";
+export { HomePage } from "./HomePage/HomePage";
+export { KingmojangPage } from "./KingmojangPage/KingmojangPage";
